+++ encode_png.h
@@ -23,7 +23,7 @@
 #include <png.h>
 #include "fill_bits.h"
 
-	void encode_png_init(struct fb_var_screeninfo sc, int fb_effective_bytes_per_pixel);
+	void encode_png_init(struct fb_var_screeninfo sc, uint8_t fb_effective_bytes_per_pixel);
 	uint8_t *encode_png(uint8_t fb_effective_bytes_per_pixel, void *fb_buf_1dim, uint32_t *imagesize);
 
 #endif /* __ENCODE_PNG_H_INCLUDED__ */
